+++ turbESN/__init__.py
@@ -1,4 +1,8 @@
+from ._version import __version__
+
 from .core import *
 from .util import *
 from .study import * 
 from .cross_validation import *
+
+